streamingccl: add application name to partition client conns
This makes it easier to quickly identify replication stream related
sessions in the output of SHOW SESSIONS.

Also by including the word "job" and the streamID in the app name, we
get a nice side-effect: SQL attaches the app name to the pprof labels
and thus these connections show up in the Advanced Debug CPU profile
for the producer side job.

Epic: none

Release note: None
